[TASK] Move vendor and public folder into dedicated .Build folder
To declutter the main directory for development, the `vendor/` and `public/`
folder are moved to a `.Build/` folder. This is good practise, for example,
used in EXT:news.

The `vendor/bin/` folder has also been moved to `.Build/bin/` to make the
access easier (would be `.Build/vendor/bin` otherwise).

Additionally, two plugins are allowed as required by Composer 2.